+++ packages/flutter/lib/src/cupertino/text_selection.dart
@@ -4,7 +4,6 @@
 
 import 'dart:collection';
 import 'dart:math' as math;
-import 'dart:ui' as ui;
 
 import 'package:flutter/widgets.dart';
 import 'package:flutter/rendering.dart';
@@ -13,6 +12,7 @@ import 'package:flutter/services.dart';
 import 'button.dart';
 import 'colors.dart';
 import 'localizations.dart';
+import 'text_selection_toolbar.dart';
 import 'theme.dart';
 
 // Read off from the output on iOS 12. This color does not vary with the
@@ -38,7 +38,6 @@ const double _kToolbarContentDistance = 8.0;
 // The height of the toolbar, including the arrow.
 const double _kToolbarHeight = 43.0;
 const Size _kToolbarArrowSize = Size(14.0, 7.0);
-const Radius _kToolbarBorderRadius = Radius.circular(8);
 // Colors extracted from https://developer.apple.com/design/resources/.
 // TODO(LongCatIsLooong): https://github.com/flutter/flutter/issues/41507.
 const Color _kToolbarBackgroundColor = Color(0xEB202020);
@@ -196,7 +195,7 @@ class _CupertinoTextSelectionToolbarWrapperState extends State<_CupertinoTextSel
       addToolbarButton(localizations.selectAllButtonLabel, widget.handleSelectAll!);
     }
 
-    return CupertinoTextSelectionToolbar._(
+    return CupertinoTextSelectionToolbar(
       barTopY: widget.barTopY,
       arrowTipX: widget.arrowTipX,
       isArrowPointingDown: widget.isArrowPointingDown,
@@ -205,212 +204,6 @@ class _CupertinoTextSelectionToolbarWrapperState extends State<_CupertinoTextSel
         children: items,
       ),
     );
-  }
-}
-
-/// An iOS-style toolbar that appears in response to text selection.
-///
-/// Typically displays buttons for text manipulation, e.g. copying and pasting text.
-///
-/// See also:
-///
-///  * [TextSelectionControls.buildToolbar], where [CupertinoTextSelectionToolbar]
-///    will be used to build an iOS-style toolbar.
-@visibleForTesting
-class CupertinoTextSelectionToolbar extends SingleChildRenderObjectWidget {
-  const CupertinoTextSelectionToolbar._({
-    Key? key,
-    required double barTopY,
-    required double arrowTipX,
-    required bool isArrowPointingDown,
-    Widget? child,
-  }) : _barTopY = barTopY,
-       _arrowTipX = arrowTipX,
-       _isArrowPointingDown = isArrowPointingDown,
-       super(key: key, child: child);
-
-  // The y-coordinate of toolbar's top edge, in global coordinate system.
-  final double _barTopY;
-
-  // The y-coordinate of the tip of the arrow, in global coordinate system.
-  final double _arrowTipX;
-
-  // Whether the arrow should point down and be attached to the bottom
-  // of the toolbar, or point up and be attached to the top of the toolbar.
-  final bool _isArrowPointingDown;
-
-  @override
-  _ToolbarRenderBox createRenderObject(BuildContext context) => _ToolbarRenderBox(_barTopY, _arrowTipX, _isArrowPointingDown, null);
-
-  @override
-  void updateRenderObject(BuildContext context, _ToolbarRenderBox renderObject) {
-    renderObject
-      ..barTopY = _barTopY
-      ..arrowTipX = _arrowTipX
-      ..isArrowPointingDown = _isArrowPointingDown;
-  }
-}
-
-class _ToolbarParentData extends BoxParentData {
-  // The x offset from the tip of the arrow to the center of the toolbar.
-  // Positive if the tip of the arrow has a larger x-coordinate than the
-  // center of the toolbar.
-  double? arrowXOffsetFromCenter;
-  @override
-  String toString() => 'offset=$offset, arrowXOffsetFromCenter=$arrowXOffsetFromCenter';
-}
-
-class _ToolbarRenderBox extends RenderShiftedBox {
-  _ToolbarRenderBox(
-    this._barTopY,
-    this._arrowTipX,
-    this._isArrowPointingDown,
-    RenderBox? child,
-  ) : super(child);
-
-
-  @override
-  bool get isRepaintBoundary => true;
-
-  double _barTopY;
-  set barTopY(double value) {
-    if (_barTopY == value) {
-      return;
-    }
-    _barTopY = value;
-    markNeedsLayout();
-    markNeedsSemanticsUpdate();
-  }
-
-  double _arrowTipX;
-  set arrowTipX(double value) {
-    if (_arrowTipX == value) {
-      return;
-    }
-    _arrowTipX = value;
-    markNeedsLayout();
-    markNeedsSemanticsUpdate();
-  }
-
-  bool _isArrowPointingDown;
-  set isArrowPointingDown(bool value) {
-    if (_isArrowPointingDown == value) {
-      return;
-    }
-    _isArrowPointingDown = value;
-    markNeedsLayout();
-    markNeedsSemanticsUpdate();
-  }
-
-  final BoxConstraints heightConstraint = const BoxConstraints.tightFor(height: _kToolbarHeight);
-
-  @override
-  void setupParentData(RenderObject child) {
-    if (child.parentData is! _ToolbarParentData) {
-      child.parentData = _ToolbarParentData();
-    }
-  }
-
-  @override
-  Size computeDryLayout(BoxConstraints constraints) {
-    return constraints.biggest;
-  }
-
-  @override
-  void performLayout() {
-    final BoxConstraints constraints = this.constraints;
-    size = constraints.biggest;
-
-    if (child == null) {
-      return;
-    }
-    final BoxConstraints enforcedConstraint = constraints
-      .deflate(const EdgeInsets.symmetric(horizontal: _kToolbarScreenPadding))
-      .loosen();
-
-    child!.layout(heightConstraint.enforce(enforcedConstraint), parentUsesSize: true,);
-    final _ToolbarParentData childParentData = child!.parentData! as _ToolbarParentData;
-
-    // The local x-coordinate of the center of the toolbar.
-    final double lowerBound = child!.size.width/2 + _kToolbarScreenPadding;
-    final double upperBound = size.width - child!.size.width/2 - _kToolbarScreenPadding;
-    final double adjustedCenterX = _arrowTipX.clamp(lowerBound, upperBound);
-
-    childParentData.offset = Offset(adjustedCenterX - child!.size.width / 2, _barTopY);
-    childParentData.arrowXOffsetFromCenter = _arrowTipX - adjustedCenterX;
-  }
-
-  // The path is described in the toolbar's coordinate system.
-  Path _clipPath() {
-    final _ToolbarParentData childParentData = child!.parentData! as _ToolbarParentData;
-    final Path rrect = Path()
-      ..addRRect(
-        RRect.fromRectAndRadius(
-          Offset(0, _isArrowPointingDown ? 0 : _kToolbarArrowSize.height)
-            & Size(child!.size.width, child!.size.height - _kToolbarArrowSize.height),
-          _kToolbarBorderRadius,
-        ),
-      );
-
-    final double arrowTipX = child!.size.width / 2 + childParentData.arrowXOffsetFromCenter!;
-
-    final double arrowBottomY = _isArrowPointingDown
-      ? child!.size.height - _kToolbarArrowSize.height
-      : _kToolbarArrowSize.height;
-
-    final double arrowTipY = _isArrowPointingDown ? child!.size.height : 0;
-
-    final Path arrow = Path()
-      ..moveTo(arrowTipX, arrowTipY)
-      ..lineTo(arrowTipX - _kToolbarArrowSize.width / 2, arrowBottomY)
-      ..lineTo(arrowTipX + _kToolbarArrowSize.width / 2, arrowBottomY)
-      ..close();
-
-    return Path.combine(PathOperation.union, rrect, arrow);
-  }
-
-  @override
-  void paint(PaintingContext context, Offset offset) {
-    if (child == null) {
-      return;
-    }
-
-    final _ToolbarParentData childParentData = child!.parentData! as _ToolbarParentData;
-    _clipPathLayer = context.pushClipPath(
-      needsCompositing,
-      offset + childParentData.offset,
-      Offset.zero & child!.size,
-      _clipPath(),
-      (PaintingContext innerContext, Offset innerOffset) => innerContext.paintChild(child!, innerOffset),
-      oldLayer: _clipPathLayer
-    );
-  }
-
-  ClipPathLayer? _clipPathLayer;
-  Paint? _debugPaint;
-
-  @override
-  void debugPaintSize(PaintingContext context, Offset offset) {
-    assert(() {
-      if (child == null) {
-        return true;
-      }
-
-      _debugPaint ??= Paint()
-        ..shader = ui.Gradient.linear(
-          const Offset(0.0, 0.0),
-          const Offset(10.0, 10.0),
-          const <Color>[Color(0x00000000), Color(0xFFFF00FF), Color(0xFFFF00FF), Color(0x00000000)],
-          const <double>[0.25, 0.25, 0.75, 0.75],
-          TileMode.repeated,
-        )
-        ..strokeWidth = 2.0
-        ..style = PaintingStyle.stroke;
-
-      final _ToolbarParentData childParentData = child!.parentData! as _ToolbarParentData;
-      context.canvas.drawPath(_clipPath().shift(offset + childParentData.offset), _debugPaint!);
-      return true;
-    }());
   }
 }
